move method tests to post_handler_test
